backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.124.0",
    "markdown-it-py>=4.0.0",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.1",
    "uvicorn>=0.38.0",
]
